Lightbulb xy fix (#170)
* HSBType.toXY() returned x y as 0-100 while the rest of the app assumes 0.0-1.0

Including HSBType’s own fromXY

* lightbulb: report hue result in hue set callback

* lightbulb: xy: fix race conditions when setting hue and saturation

Lightbulbs with xy support turn hue, saturation and brightness into an xy value to send to the lightbulb. Homebridge / HomeKit sends hue and saturation changes as separate requests. The previous implementation suffered from a race condition as the separate hue and saturation set requests were not aware of each other, wouldn’t see the change each other had made due to timing as the result was only stored once it had succeeded, and would therefore overwrite each others changes.

This change standardises the approach to get and set in the xy implementation. It always updates internal state _before_ sending the instruction to the bulb. It always uses its internal state for building a new setColorXY request, so no set requests are lost. And it uses setColorXY instead of setColorRGB, now that the result of HSBType.toXY() has been corrected (see 9489747a6806a9bd7b302a076277bcb52ecc7959).

This completely corrects the behaviour of my Philips Hue bulbs. Before this change, asking Siri to make a change would >50% of the time result in a flash to the new state then back to the old (as explained above). Now the changes work 100% of the time.